# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Agents Management Platform API
  description: |
    Enterprise-grade API for managing AI agents, executions, tools, and analytics.
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <your_access_token>
    ```
    
    ## Rate Limiting
    - API endpoints: 100 requests per minute
    - Auth endpoints: 5 requests per minute
    
    ## WebSocket
    Real-time updates available at `/socket.io` for run streaming and notifications.
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server
...
paths:
...
  /tools:
    get:
      tags: [Tools]
      summary: List tools
      operationId: listTools
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'

    post:
      tags: [Tools]
      summary: Create tool
      operationId: createTool
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateToolDto'
      responses:
        '201':
          description: Tool created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'

  # ==================== ANALYTICS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Tool:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
        schema:
          type: object
        isActive:
          type: boolean
    CreateToolDto:
      type: object
      required: [name, type, schema]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        schema:
          type: object
        implementation:
          type: string

    # Analytics
...
security:
  - bearerAuth: []